+++ jmop-src/jmop-player-fascade/.gitignore
@@ -1,0 +1,1 @@
+/target/
