+++ jmop-src/jmop-player-fascade/pom.xml
@@ -1,0 +1,32 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+	<parent>
+		<groupId>cz.martlin</groupId>
+		<artifactId>jmop</artifactId>
+		<version>3.0.1</version>
+	</parent>
+	
+	<artifactId>jmop-player-fascade</artifactId>
+	<name>The JMOP player fascade</name>
+	<description>The encapsulation of all the JMOP-player internals for the usage of the (G)UI</description>
+	
+	<dependencies>
+		<dependency>
+			<groupId>cz.martlin</groupId>
+			<artifactId>jmop-player-base</artifactId>
+			<version>3.0.1</version>
+		</dependency>
+		<dependency>
+			<groupId>cz.martlin</groupId>
+			<artifactId>jmop-player-players</artifactId>
+			<version>3.0.1</version>
+		</dependency>
+		<dependency>
+			<groupId>cz.martlin</groupId>
+			<artifactId>jmop-player-engine</artifactId>
+			<version>3.0.1</version>
+		</dependency>
+	</dependencies>
+</project>
