--- jmop-src/jmop-player-players/src/test/java/cz/martlin/jmop/core/player/PlaylistTest.java
@@ -1,12 +1,0 @@
-package cz.martlin.jmop.core.player;
-
-import org.junit.jupiter.api.Test;
-
-public class PlaylistTest {
-
-	@Test
-	public void test() {
-		// TODO ?
-	}
-
-}
