+++ jmop-src/jmop-player-players/src/test/java/cz/martlin/jmop/player/players/PlaylistTest.java
@@ -1,0 +1,12 @@
+package cz.martlin.jmop.player.players;
+
+import org.junit.jupiter.api.Test;
+
+public class PlaylistTest {
+
+	@Test
+	public void test() {
+		// TODO ?
+	}
+
+}
